+++ tests/helpers/test_node_is_pytest_context_manager.py
@@ -3,19 +3,34 @@ import ast
 import asttokens
 import pytest
 
-from flake8_aaa.helpers import node_is_pytest_raises
+from flake8_aaa.helpers import node_is_pytest_context_manager
 
 
 @pytest.mark.parametrize(
-    'code_str', [
+    'code_str',
+    [
+        # raises() with no vars
         '''
 def test():
     with pytest.raises(Exception):
         do_thing()
 ''',
+        # raises() with excinfo var
         '''
 def test_other():
     with pytest.raises(Exception) as excinfo:
+        do_thing()
+''',
+        # deprecated_call()
+        '''
+def test_api2_deprecated() -> None:
+    with pytest.deprecated_call():
+        api_call_v2()
+''',
+        # warns()
+        '''
+def test_thing_warning() -> None:
+    with pytest.warns(RuntimeWarning):
         do_thing()
 ''',
     ]
@@ -23,7 +38,7 @@ def test_other():
 def test(first_node_with_tokens):
     with_node = first_node_with_tokens.body[0]
 
-    result = node_is_pytest_raises(with_node)
+    result = node_is_pytest_context_manager(with_node)
 
     assert result is True
 
@@ -37,6 +52,6 @@ def test_no(code_str):
     asttokens.ASTTokens(code_str, tree=tree)
     node = tree.body[0]
 
-    result = node_is_pytest_raises(node)
+    result = node_is_pytest_context_manager(node)
 
     assert result is False